small refactor of main.py, add requirements.txt
