+++ .gitignore
@@ -7,6 +7,7 @@ __pycache__/
 venv
 .env
 env
+dev-venv
 
 # Unittest reports
 .coverage*
